extension/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
httpx
beautifulsoup4